move common code into lib
